warning.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@6827 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/linux/server.cpp
@@ -940,7 +940,7 @@ int main (int argc, char **argv)
         }
         if (gAutoShutdown)
             apszArgs[i++] = "--auto-shutdown";
-        apszArgs[i++] = NULL; Assert(i <= RT_ELEMENTS(apszArgs));
+        apszArgs[i++] = NULL; Assert(i <= (int)RT_ELEMENTS(apszArgs));
         execv (apszArgs[0], (char * const *)apszArgs);
         exit (0);
 # endif
